chore(chapter10): rename case02.py and test_case02.py, and update imports; add new case03_2.py
