+++ chapter10/src/case02/case02_1.py
