+++ chapter10/src/case02/case02_2.py
